AI Logic still in progress
Added a crappy solution, whereby if the ai has no cards of the
appropriate suit, it will play one of its cards at random.  This was
causing the bizzare memory error from earlier.  It no longer seg-faults,
but it does mean that it will play sub-optimally until I get that fixed,
along with the rest of the logic.  At least it sort of works.

NOTE:  Still need to check that the game is actually functioning
properly, and re-add the check of move legality.

Next, I am planning to add memory freeing functions, since the idea is
to have a daemon that can manage multiple games simultaneously.