Fixed seg fault when ai did not have led suit
The AI was segfaulting because the initial value of the lowest suit
would be 0, i.e. hearts, meaning if it did not have hearts, it would
always try to play that if it did not have any cards of the suit led,
which meant getting a non-existent card, which would lead to a segfault
the next round.  Now, it starts the suit at -1 and does not pick a suit
as weakest unless it has at least one card of that suit.