AI and move legality
AI will now play the weakest card in its hand of the suit chosen if any
previous player has played a trump card, or if any previouse player has
played a stronger card of the suit led, or if it would be playing
the winning card against an ally.

This works by first checking for any stronger cards in the trick,
including trumps.  If present, play weakest.  Otherwise, we have the
strongest card and can win the trick, so only do that if