Some little progress on the server
Did some work on the server side of things.  Need to implement a way to
push messages to a pool, then pull them off of it properly, since the
current way is not working.  Some messages get passed on properly, but
the one that is supposed to list cards seems to do nothing.